sdk_to_mcp/skills/sdk_to_mcp/pyproject.toml
[project]
name = "sdk-to-mcp"
version = "0.1.0"
description = "将硬件 SDK/协议文档自动转换为标准化 MCP Server"
readme = "SKILL.md"
requires-python = ">=3.10"
license = {text = "MIT"}

authors = [
    {name = "ROSClaw Team"}
]

dependencies = [
    "mcp[cli]>=1.13.0",
    "fastmcp>=2.11.3",
    "pypdf>=4.0.0",
    "jinja2>=3.1.0",
    "pydantic>=2.0.0",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "ruff>=0.1.0",
]
llm = [
    "litellm>=1.0.0",
]
serial = [
    "pyserial>=3.5",
]
ros = [
    "rclpy>=3.3.0",
]

[project.scripts]
sdk-to-mcp = "main:main"

[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
py-modules = ["main"]

# Ruff configuration
[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "I"]
ignore = ["E501"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"

# Pytest configuration
[tool.pytest.ini_options]
testpaths = ["."]
python_files = "test_*.py"
python_functions = "test_*"
addopts = "-v --tb=short"
asyncio_mode = "auto"
